+++ tests/test_wsgi.py
@@ -23,6 +23,7 @@ from werkzeug._compat import BytesIO, NativeStringIO, StringIO, to_bytes, \
 from werkzeug.exceptions import BadRequest, ClientDisconnected
 from werkzeug.test import Client, create_environ, run_wsgi_app
 from werkzeug.wrappers import BaseResponse
+from werkzeug.urls import url_parse
 from werkzeug.wsgi import _RangeWrapper, wrap_file
 
 
@@ -468,3 +469,51 @@ def test_range_wrapper():
         assert next(range_wrapper) == b'UND\n'
         with pytest.raises(StopIteration):
             next(range_wrapper)
+
+
+def test_http_proxy(dev_server):
+    APP_TEMPLATE = r'''
+    from werkzeug.wrappers import Request, Response
+
+    @Request.application
+    def app(request):
+        return Response(u'%s|%s|%s' % (
+            request.headers.get('X-Special'),
+            request.environ['HTTP_HOST'],
+            request.path,
+        ))
+    '''
+
+    server = dev_server(APP_TEMPLATE)
+
+    app = wsgi.ProxyMiddleware(BaseResponse('ROOT'), {
+        '/foo': {
+            'target': server.url,
+            'host': 'faked.invalid',
+            'headers': {'X-Special': 'foo'},
+        },
+        '/bar': {
+            'target': server.url,
+            'host': None,
+            'remove_prefix': True,
+            'headers': {'X-Special': 'bar'},
+        },
+        '/autohost': {
+            'target': server.url,
+        },
+    })
+
+    client = Client(app, response_wrapper=BaseResponse)
+
+    rv = client.get('/')
+    assert rv.data == b'ROOT'
+
+    rv = client.get('/foo/bar')
+    assert rv.data.decode('ascii') == 'foo|faked.invalid|/foo/bar'
+
+    rv = client.get('/bar/baz')
+    assert rv.data.decode('ascii') == 'bar|localhost|/baz'
+
+    rv = client.get('/autohost/aha')
+    assert rv.data.decode('ascii') == 'None|%s|/autohost/aha' % url_parse(
+        server.url).ascii_host